Add Dockerfile
Build a Docker image containing the statically-linked Rust binary. The
binary is created by running a muslrust container as it compiles openssl
and links it against musl. The application image is based on a scratch
image containing just that binary.

Provided a Makefile to facilitate compiling the binary, building the image
and running the container.